scripts/requirements.txt
insightface
onnxruntime
numpy
scikit-learn
Pillow
tqdm
opencv-python
supabase
